moving code out of init
